fact_sales.sql

+++ Design Data Mart/fact_sales.sql
@@ -2,10 +2,7 @@ USE EO_AdventureWorksDW2019
 
 go
 
-IF EXISTS (SELECT *
-           FROM   sys.tables
-           WHERE  NAME = 'fact_sales')
-  DROP TABLE fact_sales;
+IF EXISTS (SELECT * FROM   sys.tables WHERE  NAME = 'fact_sales') DROP TABLE fact_sales;
 
 CREATE TABLE fact_sales
   (
@@ -32,38 +29,22 @@ CREATE TABLE fact_sales
      CONSTRAINT fk_fact_sales_dim_date FOREIGN KEY (order_date_key) REFERENCES dim_date(date_key));
 
 -- Create Indexes
-IF EXISTS (SELECT *
-           FROM   sys.indexes
-           WHERE  NAME = 'fact_sales_dim_product'
-                  AND object_id = Object_id('fact_sales'))
+IF EXISTS (SELECT * FROM   sys.indexes WHERE  NAME = 'fact_sales_dim_product' AND object_id = Object_id('fact_sales'))
   DROP INDEX fact_sales.fact_sales_dim_product;
 
-CREATE INDEX fact_sales_dim_product
-  ON fact_sales(product_key);
+CREATE INDEX fact_sales_dim_product ON fact_sales(product_key);
 
-IF EXISTS (SELECT *
-           FROM   sys.indexes
-           WHERE  NAME = 'fact_sales_dim_customer'
-                  AND object_id = Object_id('fact_sales'))
+IF EXISTS (SELECT * FROM   sys.indexes WHERE  NAME = 'fact_sales_dim_customer' AND object_id = Object_id('fact_sales'))
   DROP INDEX fact_sales.fact_sales_dim_customer;
 
-CREATE INDEX fact_sales_dim_customer
-  ON fact_sales(customer_key);
+CREATE INDEX fact_sales_dim_customer ON fact_sales(customer_key);
 
-IF EXISTS (SELECT *
-           FROM   sys.indexes
-           WHERE  NAME = 'fact_sales_dim_territory'
-                  AND object_id = Object_id('fact_sales'))
+IF EXISTS (SELECT * FROM   sys.indexes WHERE  NAME = 'fact_sales_dim_territory' AND object_id = Object_id('fact_sales'))
   DROP INDEX fact_sales.fact_sales_dim_territory;
 
-CREATE INDEX fact_sales_dim_territory
-  ON fact_sales(territory_key);
+CREATE INDEX fact_sales_dim_territory ON fact_sales(territory_key);
 
-IF EXISTS (SELECT *
-           FROM   sys.indexes
-           WHERE  NAME = 'fact_sales_dim_date'
-                  AND object_id = Object_id('fact_sales'))
+IF EXISTS (SELECT * FROM   sys.indexes WHERE  NAME = 'fact_sales_dim_date' AND object_id = Object_id('fact_sales'))
   DROP INDEX fact_sales.fact_sales_dim_date;
 
-CREATE INDEX fact_sales_dim_date
-  ON fact_sales(order_date_key); 
+CREATE INDEX fact_sales_dim_date ON fact_sales(order_date_key);
